Compilation issues in 32-bit machines. Avoiding use of 'size_t' or 'long'.

+++ src/arch/mips/asm/Asm.cc
@@ -276,7 +276,7 @@ void Asm::DisassembleBinary(string path)
 	Inst inst(this);
 
 	unsigned int curr_sym;
-	long pos;
+	unsigned int pos;
 
 	/* Read Sections */
 	for (auto it = file.section_list.begin();
